fix up server restart session logic
